+++ flixel/src/org/flixel/system/debug/Perf.java
@@ -48,7 +48,7 @@ public class Perf extends FlxWindow
 	 * @param BGColor		What color the window background should be, default is gray and transparent.
 	 * @param TopColor		What color the window header bar should be, default is black and transparent.
 	 */
-	public Perf(String Title, float Width, float Height, boolean Resizable, Rectangle Bounds, long BGColor, long TopColor)
+	public Perf(String Title, float Width, float Height, boolean Resizable, Rectangle Bounds, int BGColor, int TopColor)
 	{
 		super(Title, Width, Height, Resizable, Bounds, BGColor, TopColor);
 		resize(90,66);
@@ -85,7 +85,7 @@ public class Perf extends FlxWindow
 	 * @param Bounds		A rectangle indicating the valid screen area for the window.
 	 * @param BGColor		What color the window background should be, default is gray and transparent.
 	 */
-	public Perf(String Title, float Width, float Height, boolean Resizable, Rectangle Bounds, long BGColor)
+	public Perf(String Title, float Width, float Height, boolean Resizable, Rectangle Bounds, int BGColor)
 	{
 		this(Title, Width, Height, Resizable, Bounds, BGColor, 0x7f000000);
 	}
